examples/langchain-rag/pyproject.toml
[tool.poetry]
name = "langchain-rag"
version = "0.1.0"
description = ""
authors = ["Jose F. Romaniello <jose@okta.com>"]
readme = "README.md"

[tool.poetry.scripts]
langchain-rag-example = "langchain_rag.main:main"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.11"
langchain-openai = "^0.2.12"
faiss-cpu = "^1.9.0.post1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
